executor: migrate test-infra to testify for part4 of executor_test
